Bug fixes and line fit tweaks that improve convergence.
Bug fixes:
1. Issued copy statements on some variable assignments, particularly in q3da and fitspec. These were causing problems in some plots.
2. 'blrpar' was being set in um rather than km/s; fixed. It was also being set incorrectly in run_q3dfit.
3. Removed all instances of float64 assignments throughout code.
4. Error in 'spectConvol' call when using multicore.
5. Error in plots when full wavelength information not present in edge spaxels.
6. 'q3da' no longers chokes when no data are present in a spaxel.
7. Incorrect range assignment in Legendre polynomial in qsohostfcn. Was remapping to (0,1) instead of (-1,1).

Line fit tweaks:
1. Added back 'x_scale': 'jac' to fit keywords.
2. Reduced maximum # of function evaluations to more moderate value. Implemented option to change via 'argslinefit'.
3. Added function to track parameter changes across iterations if needed.
4. Added status tracking.

+++ q3dfit/readcube.py
@@ -118,14 +118,14 @@ class Cube:
 
         # Data
         try:
-            self.dat = np.array((hdu[datext].data).T, dtype='float64')
+            self.dat = np.array((hdu[datext].data).T, dtype='float32')
         except (IndexError or KeyError):
             raise CubeError('Data extension not properly specified or absent')
 
         # Variance/Error
         if varext is not None:
             try:
-                self.var = np.array((hdu[varext].data).T, dtype='float64')
+                self.var = np.array((hdu[varext].data).T, dtype='float32')
             except (IndexError or KeyError):
                 raise CubeError('Variance extension not properly specified ' +
                                 'or absent')
@@ -324,7 +324,7 @@ class Cube:
 
         # Flux unit conversions
         # default working flux unit is erg/s/cm^2/um/sr or erg/s/cm^2/um
-        convert_flux = np.float64(1.)
+        convert_flux = np.float32(1.)
         if u.Unit(self.fluxunit_in) == u.Unit('MJy/sr') or \
             u.Unit(self.fluxunit_in) == u.Unit('MJy'):
             # IR units: https://coolwiki.ipac.caltech.edu/index.php/Units
@@ -344,7 +344,7 @@ class Cube:
                 u.Unit('erg/s/cm2/Angstrom/arcsec2/sr') or \
             u.Unit(self.fluxunit_in) == \
                 u.Unit('erg/s/cm2/Angstrom/arcsec2'):
-            convert_flux = np.float64(1e4)
+            convert_flux = np.float32(1e4)
 
         # case of output flux units: erg/s/cm^2/Anstrom (/arcsec2/sr)
         if u.Unit(self.fluxunit_out) == u.Unit('erg/s/cm2/Angstrom/sr') or \
@@ -353,7 +353,7 @@ class Cube:
                 u.Unit('erg/s/cm2/Angstrom/arcsec2/sr') or \
             u.Unit(self.fluxunit_out) == \
                 u.Unit('erg/s/cm2/Angstrom/arcsec2'):
-            convert_flux /= np.float64(1e4)
+            convert_flux /= np.float32(1e4)
 
         # self.dat = self.dat * convert_flux
         # self.var = self.var * convert_flux**2
